requirements.txt
colorama==0.4.6
PyMySQL==1.0.3
PyYAML==6.0
requests==2.31.0
SQLAlchemy==2.0.15
xmltodict==0.13.0
